+++ tests/testthat/tests.priors.R
@@ -139,3 +139,23 @@ test_that("as.brmsprior works correctly", {
   expect_equal(bprior$x, NULL)
   expect_equal(bprior$lb, rep(NA_character_, 2))
 })
+
+test_that("prior tags are correctly applied", {
+
+  prior1 <- prior(normal(0, 1), class = sd, tag = "prior_tag1")
+  prior2 <- prior(normal(0, 5), class = b, tag = "prior_tag2")
+  prior3 <- prior(normal(0, 0.5), coef = "Trt1", tag = "prior_tag3")
+  prior4 <- prior(normal(0, 10), class = "Intercept", tag = "prior_tag4")
+  prior5 <- prior(lkj_corr_cholesky(3), class = "L", group = "visit", tag = "prior_tag5")
+
+  v <- validate_prior(
+    c(prior1, prior2, prior3, prior4, prior5),
+    formula = count ~ zBase * Trt + (1 | patient) + (1 + Trt | visit),
+    data = epilepsy, family = poisson())
+
+  expect_equal(v[which(v$class == "sd"),]$tag[[1]], "prior_tag1")
+  expect_equal(v[which(v$class == "b" & v$coef != "Trt1"),]$tag[[1]], "prior_tag2")
+  expect_equal(v[which(v$class == "b" & v$coef == "Trt1"),]$tag, "prior_tag3")
+  expect_equal(v[which(v$class == "Intercept"),]$tag, "prior_tag4")
+  expect_equal(v[which(v$class == "L"),]$tag[[2]], "prior_tag5")
+})